@@ -39,48 +41,48 @@
     }
   ],
   "dependencies": {
-    "follow-redirects": "1.5.10"
+    "follow-redirects": "^1.14.4"
   },
   "deprecated": false,
   "description": "Promise based HTTP client for the browser and node.js",
   "devDependencies": {
-    "bundlesize": "^0.17.0",
+    "abortcontroller-polyfill": "^1.5.0",
     "coveralls": "^3.0.0",
+    "dtslint": "^4.1.6",
     "es6-promise": "^4.2.4",
-    "grunt": "^1.0.2",
+    "grunt": "^1.3.0",
     "grunt-banner": "^0.6.0",
     "grunt-cli": "^1.2.0",
     "grunt-contrib-clean": "^1.1.0",
     "grunt-contrib-watch": "^1.0.0",
-    "grunt-eslint": "^20.1.0",
-    "grunt-karma": "^2.0.0",
+    "grunt-eslint": "^23.0.0",
+    "grunt-karma": "^4.0.0",
     "grunt-mocha-test": "^0.13.3",
-    "grunt-ts": "^6.0.0-beta.19",
-    "grunt-webpack": "^1.0.18",
+    "grunt-webpack": "^4.0.2",
     "istanbul-instrumenter-loader": "^1.0.0",
     "jasmine-core": "^2.4.1",
-    "karma": "^1.3.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage": "^1.1.1",
-    "karma-firefox-launcher": "^1.1.0",
+    "karma": "^6.3.2",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-firefox-launcher": "^2.1.0",
     "karma-jasmine": "^1.1.1",
     "karma-jasmine-ajax": "^0.1.13",
-    "karma-opera-launcher": "^1.0.0",
     "karma-safari-launcher": "^1.0.0",
-    "karma-sauce-launcher": "^1.2.0",
+    "karma-sauce-launcher": "^4.3.6",
     "karma-sinon": "^1.0.5",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^1.7.0",
+    "karma-sourcemap-loader": "^0.3.8",
+    "karma-webpack": "^4.0.2",
     "load-grunt-tasks": "^3.5.2",
     "minimist": "^1.2.0",
-    "mocha": "^5.2.0",
+    "mocha": "^8.2.1",
     "sinon": "^4.5.0",
-    "typescript": "^2.8.1",
+    "terser-webpack-plugin": "^4.2.3",
+    "typescript": "^4.0.5",
     "url-search-params": "^0.10.0",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1"
+    "webpack": "^4.44.2",
+    "webpack-dev-server": "^3.11.0"
   },
-  "homepage": "https://github.com/axios/axios",
+  "homepage": "https://axios-http.com",
+  "jsdelivr": "dist/axios.min.js",
   "keywords": [
     "xhr",
     "http",
