@@ -1,29 +1,27 @@
 {
-  "_from": "follow-redirects@1.5.10",
-  "_id": "follow-redirects@1.5.10",
+  "_from": "follow-redirects@^1.14.4",
+  "_id": "follow-redirects@1.14.4",
   "_inBundle": false,
-  "_integrity": "sha512-0V5l4Cizzvqt5D44aTXbFZz+FtyXV1vrDN6qrelxtfYQKW0KO0W2T/hkE8xvGa/540LkZlkaUjO4ailYTFtHVQ==",
+  "_integrity": "sha512-zwGkiSXC1MUJG/qmeIFH2HBJx9u0V46QGUe3YR1fXG8bXQxq7fLj0RjLZQ5nubr9qNJUZrH+xUcwXEoXNpfS+g==",
   "_location": "/follow-redirects",
-  "_phantomChildren": {
-    "ms": "2.0.0"
-  },
+  "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "follow-redirects@1.5.10",
+    "raw": "follow-redirects@^1.14.4",
     "name": "follow-redirects",
     "escapedName": "follow-redirects",
-    "rawSpec": "1.5.10",
+    "rawSpec": "^1.14.4",
     "saveSpec": null,
-    "fetchSpec": "1.5.10"
+    "fetchSpec": "^1.14.4"
   },
   "_requiredBy": [
     "/axios"
   ],
-  "_resolved": "https://registry.npmjs.org/follow-redirects/-/follow-redirects-1.5.10.tgz",
-  "_shasum": "7b7a9f9aea2fdff36786a94ff643ed07f4ff5e2a",
-  "_spec": "follow-redirects@1.5.10",
-  "_where": "/home/abraham/Dev/Angular/Curso/SpotifyTokenGenerator/node_modules/axios",
+  "_resolved": "https://registry.npmjs.org/follow-redirects/-/follow-redirects-1.14.4.tgz",
+  "_shasum": "838fdf48a8bbdd79e52ee51fb1c94e3ed98b9379",
+  "_spec": "follow-redirects@^1.14.4",
+  "_where": "/home/abraham/Dev/ruta-front/SpotifyTokenGenerator/node_modules/axios",
   "author": {
     "name": "Ruben Verborgh",
     "email": "ruben@verborgh.org",
@@ -44,27 +42,27 @@
       "email": "james@talmage.io"
     }
   ],
-  "dependencies": {
-    "debug": "=3.1.0"
-  },
   "deprecated": false,
   "description": "HTTP and HTTPS modules that follow redirects.",
   "devDependencies": {
-    "concat-stream": "^1.6.0",
-    "coveralls": "^3.0.2",
-    "eslint": "^4.19.1",
-    "express": "^4.16.2",
-    "mocha": "^5.0.0",
-    "nyc": "^11.8.0"
+    "concat-stream": "^2.0.0",
+    "eslint": "^5.16.0",
+    "express": "^4.16.4",
+    "lolex": "^3.1.0",
+    "mocha": "^6.0.2",
+    "nyc": "^14.1.1"
   },
   "engines": {
     "node": ">=4.0"
   },
   "files": [
-    "index.js",
-    "create.js",
-    "http.js",
-    "https.js"
+    "*.js"
+  ],
+  "funding": [
+    {
+      "type": "individual",
+      "url": "https://github.com/sponsors/RubenVerborgh"
+    }
   ],
   "homepage": "https://github.com/follow-redirects/follow-redirects",
   "keywords": [
@@ -79,11 +77,10 @@
   "license": "MIT",
   "main": "index.js",
   "name": "follow-redirects",
-  "nyc": {
-    "reporter": [
-      "lcov",
-      "text"
-    ]
+  "peerDependenciesMeta": {
+    "debug": {
+      "optional": true
+    }
   },
   "repository": {
     "type": "git",
@@ -94,5 +91,5 @@
     "mocha": "nyc mocha",
     "test": "npm run lint && npm run mocha"
   },
-  "version": "1.5.10"
+  "version": "1.14.4"
 }
