@@ -1,27 +1,27 @@
 {
-  "_from": "mime-db@1.44.0",
-  "_id": "mime-db@1.44.0",
+  "_from": "mime-db@1.50.0",
+  "_id": "mime-db@1.50.0",
   "_inBundle": false,
-  "_integrity": "sha512-/NOTfLrsPBVeH7YtFPgsVWveuL+4SjjYxaQ1xtM1KMFj7HdxlBlxeyNLzhyJVx7r4rZGJAZ/6lkKCitSc/Nmpg==",
+  "_integrity": "sha512-9tMZCDlYHqeERXEHO9f/hKfNXhre5dK2eE/krIvUjZbS2KPcqGDfNShIWS1uW9XOTKQKqK6qbeOci18rbfW77A==",
   "_location": "/mime-db",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "mime-db@1.44.0",
+    "raw": "mime-db@1.50.0",
     "name": "mime-db",
     "escapedName": "mime-db",
-    "rawSpec": "1.44.0",
+    "rawSpec": "1.50.0",
     "saveSpec": null,
-    "fetchSpec": "1.44.0"
+    "fetchSpec": "1.50.0"
   },
   "_requiredBy": [
     "/mime-types"
   ],
-  "_resolved": "https://registry.npmjs.org/mime-db/-/mime-db-1.44.0.tgz",
-  "_shasum": "fa11c5eb0aca1334b4233cb4d52f10c5a6272f92",
-  "_spec": "mime-db@1.44.0",
-  "_where": "/home/abraham/Dev/Angular/Curso/SpotifyTokenGenerator/node_modules/mime-types",
+  "_resolved": "https://registry.npmjs.org/mime-db/-/mime-db-1.50.0.tgz",
+  "_shasum": "abd4ac94e98d3c0e185016c67ab45d5fde40c11f",
+  "_spec": "mime-db@1.50.0",
+  "_where": "/home/abraham/Dev/ruta-front/SpotifyTokenGenerator/node_modules/mime-types",
   "bugs": {
     "url": "https://github.com/jshttp/mime-db/issues"
   },
@@ -48,17 +48,17 @@
     "bluebird": "3.7.2",
     "co": "4.6.0",
     "cogent": "1.0.1",
-    "csv-parse": "4.8.9",
-    "eslint": "6.8.0",
-    "eslint-config-standard": "14.1.1",
-    "eslint-plugin-import": "2.20.2",
-    "eslint-plugin-markdown": "1.0.2",
+    "csv-parse": "4.16.3",
+    "eslint": "7.32.0",
+    "eslint-config-standard": "15.0.1",
+    "eslint-plugin-import": "2.24.2",
+    "eslint-plugin-markdown": "2.2.1",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-standard": "4.0.1",
+    "eslint-plugin-promise": "5.1.0",
+    "eslint-plugin-standard": "4.1.0",
     "gnode": "0.1.2",
-    "mocha": "7.1.1",
-    "nyc": "15.0.1",
+    "mocha": "9.1.1",
+    "nyc": "15.1.0",
     "raw-body": "2.4.1",
     "stream-to-array": "2.3.0"
   },
@@ -91,12 +91,12 @@
   "scripts": {
     "build": "node scripts/build",
     "fetch": "node scripts/fetch-apache && gnode scripts/fetch-iana && node scripts/fetch-nginx",
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
     "test-cov": "nyc --reporter=html --reporter=text npm test",
-    "test-travis": "nyc --reporter=text npm test",
     "update": "npm run fetch && npm run build",
     "version": "node scripts/version-history.js && git add HISTORY.md"
   },
-  "version": "1.44.0"
+  "version": "1.50.0"
 }
