@@ -1,8 +1,8 @@
 {
   "_from": "mime-types@~2.1.24",
-  "_id": "mime-types@2.1.27",
+  "_id": "mime-types@2.1.33",
   "_inBundle": false,
-  "_integrity": "sha512-JIhqnCasI9yD+SsmkquHBxTSEuZdQX5BuQnS2Vc7puQQQ+8yiP5AY5uWhpdv4YL4VM5c6iliiYWPgJ/nJQLp7w==",
+  "_integrity": "sha512-plLElXp7pRDd0bNZHw+nMd52vRYjLwQjygaNg7ddJ2uJtTlmnTCjWuPKxVu6//AdaRuME84SvLW91sIkBqGT0g==",
   "_location": "/mime-types",
   "_phantomChildren": {},
   "_requested": {
@@ -19,10 +19,10 @@
     "/accepts",
     "/type-is"
   ],
-  "_resolved": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.27.tgz",
-  "_shasum": "47949f98e279ea53119f5722e0f34e529bec009f",
+  "_resolved": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.33.tgz",
+  "_shasum": "1fa12a904472fafd068e48d9e8401f74d3f70edb",
   "_spec": "mime-types@~2.1.24",
-  "_where": "/home/abraham/Dev/Angular/Curso/SpotifyTokenGenerator/node_modules/accepts",
+  "_where": "/home/abraham/Dev/ruta-front/SpotifyTokenGenerator/node_modules/accepts",
   "bugs": {
     "url": "https://github.com/jshttp/mime-types/issues"
   },
@@ -44,20 +44,20 @@
     }
   ],
   "dependencies": {
-    "mime-db": "1.44.0"
+    "mime-db": "1.50.0"
   },
   "deprecated": false,
   "description": "The ultimate javascript content-type utility.",
   "devDependencies": {
-    "eslint": "6.8.0",
+    "eslint": "7.32.0",
     "eslint-config-standard": "14.1.1",
-    "eslint-plugin-import": "2.20.2",
-    "eslint-plugin-markdown": "1.0.2",
+    "eslint-plugin-import": "2.24.2",
+    "eslint-plugin-markdown": "2.2.1",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-standard": "4.0.1",
-    "mocha": "7.1.1",
-    "nyc": "15.0.1"
+    "eslint-plugin-promise": "5.1.0",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "9.1.2",
+    "nyc": "15.1.0"
   },
   "engines": {
     "node": ">= 0.6"
@@ -79,10 +79,10 @@
     "url": "git+https://github.com/jshttp/mime-types.git"
   },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "mocha --reporter spec test/test.js",
-    "test-cov": "nyc --reporter=html --reporter=text npm test",
-    "test-travis": "nyc --reporter=text npm test"
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
   },
-  "version": "2.1.27"
+  "version": "2.1.33"
 }
