@@ -1,8 +1,8 @@
 {
   "_from": "proxy-addr@~2.0.5",
-  "_id": "proxy-addr@2.0.6",
+  "_id": "proxy-addr@2.0.7",
   "_inBundle": false,
-  "_integrity": "sha512-dh/frvCBVmSsDYzw6n926jv974gddhkFPfiN8hPOi30Wax25QZyZEGveluCgliBnqmuM+UJmBErbAUFIoDbjOw==",
+  "_integrity": "sha512-llQsMLSUDUPT44jdrU/O37qlnifitDP+ZwrmmZcoSKyLKvtZxpyV0n2/bD/N4tBAAZ/gJEdZU7KMraoK1+XYAg==",
   "_location": "/proxy-addr",
   "_phantomChildren": {},
   "_requested": {
@@ -18,10 +18,10 @@
   "_requiredBy": [
     "/express"
   ],
-  "_resolved": "https://registry.npmjs.org/proxy-addr/-/proxy-addr-2.0.6.tgz",
-  "_shasum": "fdc2336505447d3f2f2c638ed272caf614bbb2bf",
+  "_resolved": "https://registry.npmjs.org/proxy-addr/-/proxy-addr-2.0.7.tgz",
+  "_shasum": "f19fe69ceab311eeb94b42e70e8c2070f9ba1025",
   "_spec": "proxy-addr@~2.0.5",
-  "_where": "/home/abraham/Dev/Angular/Curso/SpotifyTokenGenerator/node_modules/express",
+  "_where": "/home/abraham/Dev/ruta-front/SpotifyTokenGenerator/node_modules/express",
   "author": {
     "name": "Douglas Christopher Wilson",
     "email": "doug@somethingdoug.com"
@@ -31,7 +31,7 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "forwarded": "~0.1.2",
+    "forwarded": "0.2.0",
     "ipaddr.js": "1.9.1"
   },
   "deprecated": false,
@@ -40,15 +40,15 @@
     "beautify-benchmark": "0.2.4",
     "benchmark": "2.1.4",
     "deep-equal": "1.0.1",
-    "eslint": "6.8.0",
-    "eslint-config-standard": "14.1.0",
-    "eslint-plugin-import": "2.20.1",
-    "eslint-plugin-markdown": "1.0.1",
-    "eslint-plugin-node": "11.0.0",
-    "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-standard": "4.0.1",
-    "mocha": "7.0.1",
-    "nyc": "15.0.0"
+    "eslint": "7.26.0",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.23.4",
+    "eslint-plugin-markdown": "2.2.0",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "4.3.1",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "8.4.0",
+    "nyc": "15.1.0"
   },
   "engines": {
     "node": ">= 0.10"
@@ -73,10 +73,10 @@
   },
   "scripts": {
     "bench": "node benchmark/index.js",
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
-    "test-cov": "nyc --reporter=text npm test",
-    "test-travis": "nyc --reporter=html --reporter=text npm test"
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
   },
-  "version": "2.0.6"
+  "version": "2.0.7"
 }
